Update Html.php
I'm import HTML generated by Froala, a WYSIWYG JavaScript editor. IT generates

tags in each

item. As a quickwin without having to parse all child tags (which is for sure a new add on as well to be able to format list items!), I just add the raw node content.