Detect Word Internal Links and add title bookmark
PhpWord is not able to read an internal link to an existing heading. 
This is because the anchor is not detected in the reader. Here I create a new (internal) "Link" when this attribute is found.
Moreover, the Headings of the Word document usually contain a "bookmark" (useful for TOC or internal link) but the reader has not created it so that the name (reference of the bookmark) is lost.