styleid also for other Elements
all other types like paragraphs and characters should also receive a styleid. We will set this like the stylename.
Otherwise it could be misleading for some converters or readers, which work with the styleid, like for example GemBox. Microsoft Word itself does also set IDs for all other elements like characters.

We had issues when setting fonts via $phpword->addFontStyle() which lead to not producing a styleid in the word\styles.xml and GemBox showed all content of the document in the same default font style.