Lock all dependencies
We must have `composer.lock` as part of our repository. Otherwise, the
tools used for development change and break things unexpectedly.

Currently, this is the case with PHPUnit 10 that introduces quite a few
breaking changes. We cannot `composer install` and hope that we will get
a set of compatible dependencies. Instead, we must lock our tools and
upgrade them only when we want/need to upgrade them specifically, in a
dedicated commit, and not part of a totally unrelated work, because it
just so happenned that PHPUnit released something.

`composer.lock` is based on the oldest PHP version that is officially
supported by PHP core. Currently, that is 8.0. This typically gives us
compatibility out of the box for 8.0, 8.1 and 8.2 with a single lock
file.

EOL PHP versions, that we unfortunately keep supporting in PHPWord, are
"the special cases", and as such they will generate a new lock files in
CI, because we cannot have a single lock files that spread over such as
long time (almost 4 years for PHP 7.1) with major breaking change in the
language in between.