Extend the Options section of the PDF Writer documentation
When I used PHPWord to write a PDF document, recently, I followed the
docs only to find that I also had to specify a PDF renderer — otherwise
the operation failed.

After a bit of research to figure out what to do to specify one, I
thought it only right to document what I learned in the official
documentation.