Extend the Options section of the PDF Writer documentation (#2642)
* Extend the Options section of the PDF Writer documentation

When I used PHPWord to write a PDF document, recently, I followed the
docs only to find that I also had to specify a PDF renderer — otherwise
the operation failed.

After a bit of research to figure out what to do to specify one, I
thought it only right to document what I learned in the official
documentation.

* Update the docs changelog detailing the PDF Writer additions

This change updates the changelog to include the enhancement made to the
PDF Writer docs that show how to specify a PDF renderer when working
with the PDF writer.

* Fixed docs

---------

Co-authored-by: Progi1984 <progi1984@gmail.com>