fix: Allow vAlign and vMerge on Style\Cell to be set to null (#2676)
* fix: Allow vAlign and vMerge on Style\Cell to be set to null (#2673)

vAlign and vMerge are initialized as `null` in every style until it is explicitly set.
Right now, it is not possible to unset them, after it has been set once.

I've added a null-check to skip the validation, based on the default
parameter value of `null`, which indicates to me that it once was intended to work like this.

I've also fixed the type-hints, which were wrong from the start.

* docs(changelog): Add note about supporting vAlign and vMerge to be unset

* test: Extend existing table test case to check if unsetting vAlign works

* test: Add test case to check if unsetting vMerge works

This should fix the reduction in test coverage because of the
new null checking code in the Setter